(PC-20410) refacto(helpers): extract function into a helper folder
