+++ src/libs/algolia/fetchAlgolia/fetchMultipleOffers/helpers/buildOfferGenreTypes.ts
