+++ src/libs/algolia/fetchAlgolia/fetchMultipleOffers/helpers/buildOfferGenreTypesValues.native.test.ts
@@ -1,4 +1,4 @@
-import { buildOfferGenreTypesValues } from 'libs/algolia/fetchAlgolia/fetchMultipleOffers/buildOfferGenreTypesValues'
+import { buildOfferGenreTypesValues } from 'libs/algolia/fetchAlgolia/fetchMultipleOffers/helpers/buildOfferGenreTypesValues'
 import { GenreTypeMapping } from 'libs/subcategories/types'
 
 const mockGenreTypeMapping: GenreTypeMapping = {
